Updated TFC Dependency. Added missing lang entries.
