--- src/main/java/tfcmetallum/objects/ModToolMaterials.java
@@ -1,17 +1,0 @@
-package tfcmetallum.objects;
-
-import net.minecraft.item.Item;
-import net.minecraftforge.common.util.EnumHelper;
-
-public class ModToolMaterials
-{
-    public static final Item.ToolMaterial ALUMINIUM = EnumHelper.addToolMaterial("tfcm_aluminium", 1, 800, 8, 5.0f, 16);
-    public static final Item.ToolMaterial MITHRIL = EnumHelper.addToolMaterial("tfcm_mithril", 1, 1600, 10, 5.0f, 12);
-    public static final Item.ToolMaterial INVAR = EnumHelper.addToolMaterial("tfcm_invar", 2, 4500, 12, 7.0f, 12);
-    public static final Item.ToolMaterial COBALT = EnumHelper.addToolMaterial("tfcm_cobalt", 3, 2200, 13, 8.5f, 14);
-    public static final Item.ToolMaterial MANYULLIN = EnumHelper.addToolMaterial("tfcm_manyullin", 4, 1000, 15, 10.0f, 18);
-    public static final Item.ToolMaterial TITANIUM = EnumHelper.addToolMaterial("tfcm_titanium", 4, 3800, 16, 11.0f, 16);
-    public static final Item.ToolMaterial OSMIUM = EnumHelper.addToolMaterial("tfcm_osmium", 4, 7000, 20, 13.0f, 24);
-    public static final Item.ToolMaterial TUNGSTEN = EnumHelper.addToolMaterial("tfcm_tungsten", 4, 6600, 18, 12.5f, 24);
-    public static final Item.ToolMaterial TUNGSTEN_STEEL = EnumHelper.addToolMaterial("tfcm_tungsten_steel", 5, 9200, 20, 15.0f, 24);
-}
